performance boost, remove a bunch of float to int casts

+++ trunk/src/headers/gx_engine.h
@@ -383,7 +383,7 @@ namespace gx_engine
     float fatan;
     float fwarn;
     int fupsample;
-    float ffuse;
+    int ffuse;
     float fthreshold;
     //float ftubemode;
